refactor(podman): extract WSL2Check class to dedicated file
