+++ extensions/podman/packages/extension/src/utils/podman-install.spec.ts
@@ -26,7 +26,7 @@ import * as extensionObj from '../extension';
 import { releaseNotes } from '../podman5.json';
 import type { InstalledPodman } from './podman-cli';
 import type { Installer, PodmanInfo, UpdateCheck } from './podman-install';
-import { getBundledPodmanVersion, PodmanInstall, WinInstaller, WSL2Check } from './podman-install';
+import { getBundledPodmanVersion, PodmanInstall, WinInstaller } from './podman-install';
 import * as podmanInstallObj from './podman-install';
 import * as utils from './util';
 
@@ -311,196 +311,6 @@ test('expect winMemory preflight check return failure result if the machine has 
   expect(result.docLinksDescription).toBeUndefined();
   expect(result.docLinks).toBeUndefined();
   expect(result.fixCommand).toBeUndefined();
-});
-
-test('expect winWSL2 preflight check return successful result if the machine has WSL2 installed and do not need to reboot', async () => {
-  vi.spyOn(extensionApi.process, 'exec').mockImplementation(command => {
-    if (command === 'powershell.exe') {
-      return Promise.resolve({
-        stdout: 'True',
-        stderr: '',
-        command: 'command',
-      });
-    } else {
-      return Promise.resolve({
-        stdout: 'blabla',
-        stderr: '',
-        command: 'command',
-      });
-    }
-  });
-
-  const winWSLCheck = new WSL2Check(extensionContext);
-  const result = await winWSLCheck.execute();
-  expect(result.successful).toBeTruthy();
-});
-
-test('expect winWSL2 preflight check return failure result if the machine has WSL2 installed but needs a reboot', async () => {
-  vi.spyOn(extensionApi.process, 'exec').mockImplementation((command, args) => {
-    if (command === 'powershell.exe') {
-      return Promise.resolve({
-        stdout: 'True',
-        stderr: '',
-        command: 'command',
-      });
-    } else {
-      return new Promise<extensionApi.RunResult>((resolve, reject) => {
-        if (args?.[0] === '-l') {
-          // eslint-disable-next-line prefer-promise-reject-errors
-          reject({
-            exitCode: -1,
-            stdout: 'random error message\nError code: Wsl/WSL_E_WSL_OPTIONAL_COMPONENT_REQUIRED',
-            stderr: '',
-            command: 'command',
-          });
-        } else {
-          // eslint-disable-next-line prefer-promise-reject-errors
-          resolve({
-            stdout: 'blabla',
-            stderr: '',
-            command: 'command',
-          });
-        }
-      });
-    }
-  });
-
-  const winWSLCheck = new WSL2Check(extensionContext);
-  const result = await winWSLCheck.execute();
-  expect(result.description).equal(
-    'WSL2 seems to be installed but the system needs to be restarted so the changes can take effect.',
-  );
-  expect(result.docLinksDescription).equal(
-    `If already restarted, call 'wsl --install --no-distribution' in a terminal.`,
-  );
-  expect(result.docLinks?.[0].url).equal('https://learn.microsoft.com/en-us/windows/wsl/install');
-});
-
-test('expect winWSL2 preflight check return successful result if the machine has WSL2 installed and the reboot check fails with a code different from WSL_E_WSL_OPTIONAL_COMPONENT_REQUIRED', async () => {
-  vi.spyOn(extensionApi.process, 'exec').mockImplementation((command, args) => {
-    if (command === 'powershell.exe') {
-      return Promise.resolve({
-        stdout: 'True',
-        stderr: '',
-        command: 'command',
-      });
-    } else {
-      return new Promise<extensionApi.RunResult>((resolve, reject) => {
-        if (args?.[0] === '-l') {
-          // eslint-disable-next-line prefer-promise-reject-errors
-          reject({
-            exitCode: -1,
-            stdout: 'random error message\nError code: Wsl/WSL_E_DEFAULT_DISTRO_NOT_FOUND',
-            stderr: '',
-            command: 'command',
-          });
-        } else {
-          // eslint-disable-next-line prefer-promise-reject-errors
-          resolve({
-            stdout: 'blabla',
-            stderr: '',
-            command: 'command',
-          });
-        }
-      });
-    }
-  });
-
-  const winWSLCheck = new WSL2Check(extensionContext);
-  const result = await winWSLCheck.execute();
-  expect(result.successful).toBeTruthy();
-});
-
-test('expect winWSL2 preflight check return failure result if user do not have wsl but he is admin', async () => {
-  vi.spyOn(extensionApi.process, 'exec').mockImplementation(command => {
-    if (command === 'powershell.exe') {
-      return Promise.resolve({
-        stdout: 'True',
-        stderr: '',
-        command: 'command',
-      });
-    } else {
-      return Promise.resolve({
-        stdout: '',
-        stderr: '',
-        command: 'command',
-      });
-    }
-  });
-
-  const winWSLCheck = new WSL2Check(extensionContext);
-  const result = await winWSLCheck.execute();
-  expect(result.description).equal('WSL2 is not installed.');
-  expect(result.docLinksDescription).equal(`Call 'wsl --install --no-distribution' in a terminal.`);
-  expect(result.docLinks?.[0].url).equal('https://learn.microsoft.com/en-us/windows/wsl/install');
-  expect(result.docLinks?.[0].title).equal('WSL2 Manual Installation Steps');
-});
-
-test('expect winWSL2 preflight check return failure result if user do not have wsl but he is not admin', async () => {
-  vi.spyOn(extensionApi.process, 'exec').mockImplementation(command => {
-    if (command === 'powershell.exe') {
-      return Promise.resolve({
-        stdout: 'False',
-        stderr: '',
-        command: 'command',
-      });
-    } else {
-      return Promise.resolve({
-        stdout: '',
-        stderr: '',
-        command: 'command',
-      });
-    }
-  });
-
-  const winWSLCheck = new WSL2Check(extensionContext);
-  const result = await winWSLCheck.execute();
-  expect(result.description).equal('WSL2 is not installed or you do not have permissions to run WSL2.');
-  expect(result.docLinksDescription).equal('Contact your Administrator to setup WSL2.');
-  expect(result.docLinks?.[0].url).equal('https://learn.microsoft.com/en-us/windows/wsl/install');
-  expect(result.docLinks?.[0].title).equal('WSL2 Manual Installation Steps');
-});
-
-test('expect winWSL2 preflight check return failure result if it fails when checking if wsl is installed', async () => {
-  vi.spyOn(extensionApi.process, 'exec').mockImplementation(command => {
-    if (command === 'powershell.exe') {
-      return Promise.resolve({
-        stdout: '',
-        stderr: '',
-        command: 'command',
-      });
-    } else {
-      throw new Error();
-    }
-  });
-
-  const winWSLCheck = new WSL2Check(extensionContext);
-  const result = await winWSLCheck.execute();
-  expect(result.description).equal('Could not detect WSL2');
-  expect(result.docLinks?.[0].url).equal('https://learn.microsoft.com/en-us/windows/wsl/install');
-  expect(result.docLinks?.[0].title).equal('WSL2 Manual Installation Steps');
-});
-
-test('expect winWSL2 init to register WSLInstall command', async () => {
-  const registerCommandMock = vi.spyOn(extensionApi.commands, 'registerCommand');
-  const winWSLCheck = new WSL2Check(extensionContext);
-  await winWSLCheck.init?.();
-  expect(registerCommandMock).toBeCalledWith('podman.onboarding.installWSL', expect.any(Function));
-});
-
-test('expect winWSL2 command to be registered as disposable', async () => {
-  const registerCommandMock = vi.spyOn(extensionApi.commands, 'registerCommand');
-  const disposableMock = {
-    dispose: vi.fn(),
-  };
-  registerCommandMock.mockReturnValue(disposableMock);
-  const winWSLCheck = new WSL2Check(extensionContext);
-  await winWSLCheck.init?.();
-  expect(registerCommandMock).toBeCalledWith('podman.onboarding.installWSL', expect.any(Function));
-
-  // should contain a subscription with a disposable function
-  expect(extensionContext.subscriptions[0]).toBeDefined();
-  expect(extensionContext.subscriptions[0].dispose).toBeDefined();
 });
 
 describe('getBundledPodmanVersion', () => {
